+++ tools/src/tester/StoryOutputs.ts
@@ -48,12 +48,13 @@ export class StoryOutputs {
   }
 
   resolve_string (str: string): any {
-    const ref = OutputReference.parse(str)
-    if (ref) {
-      return this.get_output_value(ref.chapter_id, ref.output_name)
-    } else {
-      return str
-    }
+    return OutputReference.replace(str, (chapter_id, output_name) => {
+      if (chapter_id === undefined || output_name === undefined) {
+        throw new Error(`Invalid output references in ${str}.`)
+      }
+
+      return this.get_output_value(chapter_id as string, output_name as string)
+    })
   }
 
   resolve_value (payload: any): any {