+++ tools/tests/tester/StoryOutputs.test.ts
@@ -55,13 +55,19 @@ test('resolve_params', () => {
     a: '${chapter_id.x}',
     b: '${chapter_id.y}',
     c: 3,
-    d: 'str'
+    d: 'str',
+    e: '${chapter_id.x}:${chapter_id.y}',
+    f: 'x=${chapter_id.x}',
+    g: '${chapter_id.y}=y'
   }
   expect(story_outputs.resolve_params(parameters)).toEqual({
     a: 1,
     b: 2,
     c: 3,
-    d: 'str'
+    d: 'str',
+    e: '1:2',
+    f: 'x=1',
+    g: '2=y'
   })
 })